+++ concrete-csprng/src/generators/aes_ctr/generic.rs
@@ -331,7 +331,7 @@ pub mod aes_ctr_generic_test {
             let original_generator =
                 AesCtrGenerator::<G>::new(k, Some(t), Some(t.increased(nc.0 * nb.0 + i)));
             let mut forked_generator = original_generator.clone();
-            let initial_output: Vec<u8> = original_generator.take(bytes_to_go as usize).collect();
+            let initial_output: Vec<u8> = original_generator.take(bytes_to_go).collect();
             let forked_output: Vec<u8> = forked_generator
                 .try_fork(nc, nb)
                 .unwrap()
